+++ src/main/java/com/secucard/connect/SecucardConnect.java
@@ -57,7 +57,7 @@ import java.util.TimerTask;
  * The entry point to the secucard API, provides resources for product operations.
  */
 public class SecucardConnect {
-  public static final String VERSION = "2.4.0";
+  public static final String VERSION = "2.5.0";
 
   protected volatile boolean isConnected;
   private Configuration configuration;
@@ -417,7 +417,7 @@ public class SecucardConnect {
         service(Payment.Secupayinvoices), service(Payment.Secupaycreditcards));
 
     loyalty = new Loyalty(service(Loyalty.Cards), service(Loyalty.Customers),
-        service(Loyalty.Merchantcards));
+        service(Loyalty.Merchantcards), service(Loyalty.CardGroups));
 
     services = new Services(service(Services.Identrequests), service(Services.Identresults));